backend/requirements.txt
fastapi
uvicorn
motor
pydantic>=2.0
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-genai
pillow
requests
email-validator
python-dotenv
pdf2image
